server/requirements.txt
django
django_rest_framework
psycopg2-binary
requests
pandas
traitlets
django-cors-headers
gunicorn